Authoritative form-control state with native integration
Give input/textarea/option/output one Rust-owned live state: dirty value flags with default mirroring, user-edit tracking, number editing buffers, option selectedness/dirtiness, output default overrides. Parser, attribute, child-list, clone, and type-change hooks keep state coherent; JS IDL delegates through host ops with no second value store. Layout, paint, FormData, scripted and scriptless submission/edit/reset consume the same state; select no-selection travels the wire as -1. Native text/select edits use internal user-edit operations and bypass author value setters.

+++ src/engine/layout/model.rs
@@ -53,7 +53,7 @@ pub struct ControlSpec {
     pub value: String,
     pub label: String,
     pub options: Vec<SelectOption>,
-    pub selected_index: usize,
+    pub selected_index: i64,
     pub placeholder: String,
     pub form_id: Option<NodeId>,
     pub background_color: Color,
